Added ldfConvert message about multihits.
The HRIBF LDF format may have data events where a parameter (channel) is set
multiple times. The meaning of this may be ambiguous, but perhaps is
expected, e.g. multi-hit TDCs. Currently the last parameter data readout will be stored. Due to this an error message
is printed for each event with this situation listing the buffer, event
number and parameters that were multiply hit.