Initial commit of a revised tree for goddess.
The goddess tree has been revised to use vectors of structures for each
type of detected particle. These include charged particle under SiData,
gammas under GamData and recoil particles under IonData. These
sturctures are stored in goddess/include/GoddessStruct.h and are
compiled into libGoddessStruct.so.

This data is fairly high level, so checks on the lower level raw data
must be performed to ensure the quality of the high level data. This
includes, but is not limited to checking claibrations and thresholds. It
is especially important that the noise peaks are cut out an dtha thte
digital and anlog systems are roughly gain matched.

This commit does not require any coincidences.