Added strip number to tree and updated resistive strip energies
I changed the way that the energies are extracted in GoddessData to include the differences between Ntype and Ptype strips. Also corrected the strip energy for the resistive strips. I also updated the config file to the correct format to deal with the resistive strips i.e. end to end gain matching offset and slope are entered in after the usual "enCal p #", strip energy calibrations are added in after "enCal resStrip #". The parameters currently in this goddess.config are wrong and need to be corrected. Also to note is the si.Pos is still not quite right and needs to be fixed. More to come soon!