Heavy changes to the tree structure (again)
I had to change a few (a lot) of stuffs to store data more efficiently. There is too much data to store to be able to keep track of everything (all the strips which fired, all the energies collected, the near and far contact for the resistive strips...) without having to wait eternity to plot a single graph.
There is a new running option which is sidetails, handling the detail level for the silicon detectors. sidetails=0 will not write anything about the silicon detectors in the output root file (probably useless but well...). sidetails=1 ill write only the sum of the energies collected by sector and by side (p or n) as well as the number of the strip which colelcted the max energy. sidetails=2 will write everything like before. I highly recommand not to use sidetails=2 for more than checking how consistent the data are sorted on a few millions of events max.
To improve the situation with the file size, I decided to reduce what is stored to the min possible. TRhe energies are stored in a vector<float> now instead of in 3 floats for dE, E1, and E2. This prevent the storing of useless millions of 0.0 when the dE and/or E2 layers do not exists for a particular telescope.
The issue is that it is not possible to plot something like dE vs. E1 directly using the vector<float> stored directly in the rootfile. To do that I implemented 2 functions that can be used in combination with the Draw command to retrieve specifically the energy sum and strip for a specific layer in a specific telescope.

si.eSumLayer(unsigned short layer, bool isNtype)
si.stripMaxLayer(unsigned short layer, bool isNtype)

First argument is the layer number (0 == dE, 1 == E1, 2 == E2)
Second argument is the side type: false == p-side (front), true == n-side (back)

It can be used using following this example:
god->Draw("si.eSumLayer(1, false):si.eSumLayer(0, false)","","colz")
to plot E1 vs dE for the front side